Move generate code to cli/crd/generate/
