+++ src/cli/crd/generate/generate.test.ts
